+++ openprocurement/agreement/cfaua/models/parameter.py
@@ -1,7 +1,7 @@
+# -*- coding: utf-8 -*-
 from schematics.exceptions import ValidationError
-from schematics.types import StringType, FloatType
-
-from openprocurement.api.models import Model
+from schematics.types import StringType
+from openprocurement.api.models import Model, DecimalType
 from openprocurement.api.roles import RolesFromCsv
 from openprocurement.agreement.core.utils import get_agreement
 
@@ -11,7 +11,7 @@ class Parameter(Model):
         serialize_when_none = False
         roles = RolesFromCsv('Parameter.csv', relative_to=__file__)
     code = StringType(required=True)
-    value = FloatType(required=True)
+    value = DecimalType(required=True)
 
     def validate_code(self, data, code):
         if isinstance(data['__parent__'], Model) and \
